move ColdMessageRoleType under RoleTypes namespace
