+++ app/components/role_types/cold_message_component.html.erb
