+++ app/views/cold_messages/new.html.erb
@@ -14,7 +14,7 @@
       <% end %>
     </div>
 
-    <%= render ColdMessageRoleTypeComponent.new(@cold_message.developer.role_type) %>
+    <%= render RoleTypes::ColdMessageComponent.new(@cold_message.developer.role_type) %>
   </div>
 
   <div class="mt-12 sm:mt-20 sm:mx-auto sm:w-full sm:max-w-xl">
